refactor: move function to tenderly utils
