+++ apps/web/src/components/tx-flow/TxInfoProvider.tsx
@@ -3,7 +3,7 @@ import { createContext, type ReactElement } from 'react'
 import { useSimulation } from '@/components/tx/security/tenderly/useSimulation'
 import { FETCH_STATUS } from '@safe-global/utils/components/tx/security/tenderly/types'
 import type { UseSimulationReturn } from '@safe-global/utils/components/tx/security/tenderly/useSimulation'
-import { getSimulationStatus } from '../tx/security/tenderly/utils'
+import { getSimulationStatus } from '@safe-global/utils/components/tx/security/tenderly/utils'
 
 type SimulationStatus = {
   isLoading: boolean
